changed prepare counter from 8 bytes to 64 in test/format.

+++ test/format/format.h
@@ -122,7 +122,7 @@ typedef struct {
 
 	WT_RAND_STATE rnd;			/* Global RNG state */
 
-	uint8_t prepare_cnt;			/* Counter for prepare */
+	uint64_t prepare_cnt;			/* Counter for prepare */
 
 	uint64_t timestamp;			/* Counter for timestamps */
